refactor plant detail components
