+++ src/features/plantas/detail/components/LecturasTab.jsx
@@ -1,0 +1,11 @@
+import styles from '../../../../styles/plant-detail.module.css';
+
+const LecturasTab = () => (
+  <div className={styles.tabContent}>
+    <h3>Historial de Lecturas</h3>
+    <p>Aquí se mostrará el historial completo de lecturas de la planta.</p>
+    {/* Implementar tabla de lecturas con paginación */}
+  </div>
+);
+
+export default LecturasTab;
